down to nothing
